Security: Removed user-controlled redirect URL
Github CodeQL detected a medium level security vulerability in this run: https://github.com/containerish/OpenRegistry/runs/5711928771
In GithubCallbackHandler, we receive a "path" query param, which we use
to redirect user back to our web interface. We anyway provide webui
redirect url via config so it's better to stick to that.

Signed-off-by: jay-dee7 <jasdeepsingh.uppal@gmail.com>