fix: Use new header key for OCI Chunked Push Min chunk size
OCI recently added support for defining a minimum size for chunks when
uploading an image with chunked uploading. The header key is
`OCI-Chunk-Min-Length` and the value for it is set in bytes.
We default to 5MB (5242880 Bytes) but it can be overridden via the
config.yaml file using `dfs.<storage_provider>.min_chunk_size`